Built Basic App
Created a basic habit tracker that allows users to add a new habit to a list and get details on tap.

The data and UI architecture for this app is in place: each habit is a struct; the list of habits is a class; the AddHabit view appears as a sheet; and the detail view appears via NavigationLink.